requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
attrs==25.3.0
branca==0.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
folium==0.19.5
geopandas==1.0.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pyogrio==0.10.0
pyproj==3.7.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
selenium==4.29.0
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xyzservices==2025.1.0